+++ lending/src/lib/config.ts
@@ -48,7 +48,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
     language: "typescript",
     openapiDocVersion: "1.0.0",
-    sdkVersion: "0.2.0",
-    genVersion: "2.390.6",
-    userAgent: "speakeasy-sdk/typescript 0.2.0 2.390.6 1.0.0 ryan-lending",
+    sdkVersion: "0.2.1",
+    genVersion: "2.390.0",
+    userAgent: "speakeasy-sdk/typescript 0.2.1 2.390.0 1.0.0 ryan-lending",
 } as const;